jar with dependencies
